+++ src/components/AddCard/CardConfirmModal.tsx
@@ -1,17 +1,11 @@
-import { useContext } from "react";
-import { useNavigate } from "react-router-dom";
+import { ChangeEvent, useContext } from "react";
 import styled from "styled-components";
-import {
-  CARD_REGISTER_SUCCESS_MESSAGE,
-  GUIDE_MESSAGE,
-  ROUTES,
-} from "../../constants/constants";
+import { GUIDE_MESSAGE } from "../../constants/constants";
 import { CardInfoContext } from "../../contexts/CardInfoContext";
-import { setCardAlias, setInitialState } from "../../reducer/cardReducer";
-import { isInvalidCardAlias } from "../../validators/validator.ts";
-
+import { setCardAlias } from "../../reducer/cardReducer";
+import { isInvalidCardAlias } from "../../validators/validator";
 import PageHeader from "../PageHeader";
-import Button from "../UIComponents/Button/Button.tsx";
+import Button from "../UIComponents/Button/Button";
 import CardPreview from "../UIComponents/CardPreview/CardPreview";
 import Input from "../UIComponents/Input/Input";
 
@@ -50,23 +44,14 @@ export default function CardConfirmModal() {
   const { state, dispatch } = useContext(CardInfoContext);
   const { cardNumber, holderName, expireDate, cardAlias } = state;
 
-  const navigate = useNavigate();
+  const handleCardAliasUpdate = (event: ChangeEvent<HTMLInputElement>) => {
+    const {
+      target: { value },
+    } = event;
 
-  const handleCardAliasUpdate = ({ target: { value } }) => {
     if (isInvalidCardAlias(value)) return;
 
     dispatch(setCardAlias({ value }));
-  };
-
-  const handleSubmitButton = () => {
-    window.alert(CARD_REGISTER_SUCCESS_MESSAGE);
-
-    const storage = JSON.parse(localStorage.getItem("cardInfo")) ?? [];
-    storage.push(state);
-    localStorage.setItem("cardInfo", JSON.stringify(storage));
-
-    dispatch(setInitialState());
-    navigate(ROUTES.HOME, { replace: true });
   };
 
   return (
@@ -97,9 +82,7 @@ export default function CardConfirmModal() {
           borderBottom={"1px solid #8b8b8b"}
         />
         {cardAlias.length > 0 ? (
-          <Button onClick={handleSubmitButton} type="submit">
-            확인
-          </Button>
+          <Button type="submit">확인</Button>
         ) : (
           <StyledCardAliasGuideMessage>
             {GUIDE_MESSAGE.VALID_CARD_ALIAS}
