+++ src/components/AddCard/CardSecurityCodeInput.tsx
@@ -1,19 +1,16 @@
-import { useContext, useState } from "react";
+import { ChangeEvent, useContext, useState } from "react";
 import { CardInfoContext } from "../../contexts/CardInfoContext";
 import styled from "styled-components";
-
 import HelpIconImage from "../../assets/images/questionMark.svg";
-
 import Input from "../UIComponents/Input/Input";
 import InputField from "../UIComponents/InputField/InputField";
-
+import { isInvalidSecurityCode } from "../../validators/validator";
+import { setSecurityCode } from "../../reducer/cardReducer";
 import {
   CARD_INFO_RULES,
   CREATE_MASKED_CHARACTERS,
   GUIDE_MESSAGE,
 } from "../../constants/constants";
-import { isInvalidSecurityCode } from "../../validators/validator.ts";
-import { setSecurityCode } from "../../reducer/cardReducer";
 
 const StyledIconContainer = styled.div`
   position: relative;
@@ -26,7 +23,7 @@ const StyledIcon = styled.img`
 `;
 
 const StyledDescription = styled.p`
-  ${(props) => !props.isOpen && `display: none;`}
+  ${(props: { isOpen: boolean }) => !props.isOpen && `display: none;`}
   position: absolute;
   top: 5px;
   right: -205px;
@@ -48,7 +45,7 @@ const StyledDescription = styled.p`
 export const SECURITY_CODE_DESCRIPTION =
   "CVV/CVC 번호는 카드 뒷 면에 있는 3자리 숫자이며 카드 보안을 위한 번호입니다.";
 
-export function HelpIcon({ description }) {
+export function HelpIcon({ description }: { description: string }) {
   const [isDescriptionOpen, setIsDescriptionOpen] = useState(false);
 
   return (
@@ -70,8 +67,10 @@ export default function CardSecurityCodeInput() {
     dispatch,
   } = useContext(CardInfoContext);
 
-  const handleSecurityCodeUpdate = (event) => {
-    const { value } = event.target;
+  const handleSecurityCodeUpdate = (event: ChangeEvent<HTMLInputElement>) => {
+    const {
+      target: { value },
+    } = event;
 
     if (Number.isNaN(value) || isInvalidSecurityCode(value)) {
       event.preventDefault();
@@ -99,7 +98,7 @@ export default function CardSecurityCodeInput() {
         width={"100%"}
         maxLength={CARD_INFO_RULES.SECURITY_CODE_LENGTH}
         required
-        onChange={(e) => handleSecurityCodeUpdate(e, "securityCode")}
+        onChange={(e) => handleSecurityCodeUpdate(e)}
         isComplete={
           securityCode.length === CARD_INFO_RULES.SECURITY_CODE_LENGTH
         }
