+++ src/components/PossessCard/index.tsx
@@ -1,0 +1,1 @@
+export { default as PossessCardItems } from "./PossessCardItems";
